⚙️ [codebase-cleanup] bridge(app): consolidate plugin and repository test fakes [step 9/45]
