+++ bridge/app/test/bridge/repositories/project_repository_test.dart
@@ -16,6 +16,7 @@ import "package:test/test.dart";
 
 import "../../helpers/fake_filesystem_api.dart";
 import "../../helpers/fake_git_cli_api.dart";
+import "../../helpers/fakes/fake_derived_bridge_plugin.dart";
 import "../../helpers/plugin_runtime_test_support.dart";
 import "../../helpers/test_database.dart";
 
@@ -1211,12 +1212,6 @@ PluginSession _session(
 /// Minimal [BridgePluginApi] fake for plugin activity evidence. Every other
 /// member throws so accidental project open/rename delegation is loud.
 class _FakeBridgePlugin() implements NativeProjectsPluginApi {
-  @override
-  Future<List<PluginQueuedPrompt>> getQueuedPrompts({required String sessionId}) async => const [];
-
-  @override
-  Future<bool> cancelQueuedPrompt({required String sessionId, required String promptId}) async => false;
-
   List<PluginProject> projectsResult = const [];
   Future<List<PluginProject>>? getProjectsFuture;
   Object? getProjectsError;
@@ -1224,6 +1219,12 @@ class _FakeBridgePlugin() implements NativeProjectsPluginApi {
   String? lastGetProjectId;
   String? lastRenameProjectId;
   int getProjectsCallCount = 0;
+
+  @override
+  String get id => "fake";
+
+  @override
+  Stream<BridgeSseEvent> get events => throw UnimplementedError();
 
   @override
   Future<List<PluginProject>> getProjects() async {
@@ -1235,162 +1236,16 @@ class _FakeBridgePlugin() implements NativeProjectsPluginApi {
   }
 
   @override
-  String get id => "fake";
-
-  @override
-  Stream<BridgeSseEvent> get events => throw UnimplementedError();
-
-  @override
-  Future<List<PluginSession>> getSessions(String projectId, {int? start, int? limit}) => throw UnimplementedError();
-
-  @override
-  Future<PluginSession> createSession({
-    required String directory,
-    required String? parentSessionId,
-    required List<PluginPromptPart> parts,
-    required String? userVisibleText,
-    required PluginSessionVariant? variant,
-    required String? agent,
-    required ({String providerID, String modelID})? model,
-  }) => throw UnimplementedError();
-
-  @override
-  Future<PluginSession> renameSession({required String sessionId, required String title}) => throw UnimplementedError();
-
-  @override
   Future<PluginProject> renameProject({required String projectId, required String name}) async {
     lastRenameProjectId = projectId;
     return projectResult;
   }
 
   @override
-  Future<void> deleteSession(String sessionId) => throw UnimplementedError();
-
-  @override
-  Future<void> archiveSession({required String sessionId}) => throw UnimplementedError();
-
-  @override
-  Future<void> deleteWorkspace({
-    required String projectId,
-    required String worktreePath,
-  }) => throw UnimplementedError();
-
-  @override
-  Future<List<PluginSession>> getChildSessions(String sessionId) => throw UnimplementedError();
-
-  @override
-  Future<Map<String, PluginSessionStatus>> getSessionStatuses() => throw UnimplementedError();
-
-  @override
-  Future<List<PluginMessageWithParts>> getSessionMessages(String sessionId) => throw UnimplementedError();
-
-  @override
-  Future<List<PluginCommand>> getCommands({required String? projectId}) async => <PluginCommand>[];
-
-  @override
-  Future<PluginSessionOptionsDiscoveryResult> getSessionOptions({
-    required String projectId,
-    required PluginSessionOptionsDiscoveryMode discoveryMode,
-  }) => throw UnimplementedError();
-
-  @override
-  Future<void> sendPrompt({
-    required String promptId,
-    required String sessionId,
-    required List<PluginPromptPart> parts,
-    required PluginSessionVariant? variant,
-    required String? agent,
-    required ({String providerID, String modelID})? model,
-  }) => throw UnimplementedError();
-
-  @override
-  Future<void> sendCommand({
-    required String promptId,
-    required String sessionId,
-    required String command,
-    required String arguments,
-    required String? userVisibleArguments,
-    required PluginSessionVariant? variant,
-    required String? agent,
-    required ({String providerID, String modelID})? model,
-  }) async {}
-
-  @override
-  Future<void> abortSession({required String sessionId}) => throw UnimplementedError();
-
-  @override
-  Future<List<PluginAgent>> getAgents({required String projectId}) => throw UnimplementedError();
-
-  @override
-  Future<List<PluginPendingPermission>> getPendingPermissions({required String sessionId}) async => [];
-
-  @override
-  Future<List<PluginPendingQuestion>> getPendingQuestions({required String sessionId}) => throw UnimplementedError();
-
-  @override
-  Future<List<PluginPendingQuestion>> getProjectQuestions({required String projectId}) => throw UnimplementedError();
-
-  @override
-  Future<void> replyToQuestion({
-    required String questionId,
-    required String sessionId,
-    required List<List<String>> answers,
-  }) => throw UnimplementedError();
-
-  @override
-  Future<void> rejectQuestion({required String questionId, required String? sessionId}) => throw UnimplementedError();
-
-  @override
-  Future<void> replyToPermission({
-    required String requestId,
-    required String sessionId,
-    required PluginPermissionReply reply,
-  }) => throw UnimplementedError();
-
-  @override
   Future<PluginProject> getProject(String projectId) async {
     lastGetProjectId = projectId;
     return projectResult;
   }
-
-  @override
-  Future<bool> healthCheck() => throw UnimplementedError();
-
-  @override
-  Future<PluginProvidersResult> getProviders({required String projectId}) => throw UnimplementedError();
-
-  @override
-  List<PluginProjectActivitySummary> getActiveSessionsSummary() => throw UnimplementedError();
-
-  @override
-  Future<void> dispose() => throw UnimplementedError();
-}
-
-/// A derive-style plugin: exposes its sessions through
-/// [BridgeDerivedProjectsPluginApi.listAllSessions], mirroring how Codex/ACP
-/// plugins are shaped — it has no project members at all, so the bridge
-/// derivation path is what's exercised.
-class _FakeDerivedPlugin(var List<PluginSession> sessions) implements BridgeDerivedProjectsPluginApi {
-  /// Points at a session directory used by the tests so the launch-folder seed
-  /// doesn't introduce an extra project the assertions don't expect.
-  String launchDir = "/tmp/proj/alpha";
-
-  /// The hint set received on the most recent [listAllSessions] call.
-  Set<String>? receivedKnownDirectories;
-  int listAllSessionsCallCount = 0;
-
-  @override
-  String get id => "codex";
-
-  @override
-  Future<List<PluginSession>> listAllSessions({required Set<String> knownDirectories}) async {
-    listAllSessionsCallCount++;
-    receivedKnownDirectories = knownDirectories;
-    return sessions;
-  }
-
-  @override
-  String get launchDirectory => launchDir;
 
   @override
   Future<PluginSessionOptionsDiscoveryResult> getSessionOptions({
@@ -1400,6 +1255,17 @@ class _FakeDerivedPlugin(var List<PluginSession> sessions) implements BridgeDeri
 
   @override
   dynamic noSuchMethod(Invocation invocation) => super.noSuchMethod(invocation);
+}
+
+/// A derive-style plugin: exposes its sessions through
+/// [BridgeDerivedProjectsPluginApi.listAllSessions], mirroring how Codex/ACP
+/// plugins are shaped — it has no project members at all, so the bridge
+/// derivation path is what's exercised.
+class _FakeDerivedPlugin(final List<PluginSession> initialSessions) extends FakeDerivedBridgePlugin {
+  this : super(id: "codex", launchDirectory: "/tmp/proj/alpha", allSessions: initialSessions);
+
+  List<PluginSession> get sessions => allSessions;
+  set sessions(List<PluginSession> value) => allSessions = value;
 }
 
 class _CountingProjectsDao({required AppDatabase database}) extends ProjectsDao {
