+++ bridge/app/test/bridge/routing/routing_test_helpers.dart
@@ -20,7 +20,6 @@ import "package:sesori_bridge/src/repositories/models/stored_session.dart";
 import "package:sesori_bridge/src/repositories/models/verified_github_login.dart";
 import "package:sesori_bridge/src/repositories/pr_source_repository.dart";
 import "package:sesori_bridge/src/repositories/pull_request_repository.dart";
-import "package:sesori_bridge/src/repositories/session_metadata_repository.dart";
 import "package:sesori_bridge/src/repositories/session_repository.dart";
 import "package:sesori_bridge/src/repositories/session_unseen_calculator.dart";
 import "package:sesori_bridge/src/repositories/session_unseen_repository.dart";
@@ -36,7 +35,12 @@ import "package:sesori_shared/sesori_shared.dart" hide PermissionReply;
 
 import "../../helpers/fake_filesystem_api.dart";
 import "../../helpers/fake_git_cli_api.dart";
+import "../../helpers/fakes/fake_bridge_plugin.dart";
+import "../../helpers/fakes/fake_repository_fakes.dart";
 import "../../helpers/single_plugin_repository_test_support.dart";
+
+export "../../helpers/fakes/fake_bridge_plugin.dart";
+export "../../helpers/fakes/fake_repository_fakes.dart";
 
 /// Builds a real [SessionUnseenService] backed by [db] for handler/router tests.
 SessionUnseenService buildTestSessionUnseenService(AppDatabase db, BridgePluginApi plugin) {
@@ -115,394 +119,6 @@ extension RequestHandlerTestMatching on RequestHandlerBase {
     String? fragment,
   })
   extractParams(RelayRequest request) => extractTargetParams(target: Uri.parse(request.path));
-}
-
-/// Hand-written fake [BridgePluginApi] used across routing handler tests.
-class FakeBridgePlugin() implements NativeProjectsPluginApi {
-  /// Seedable queue returned by [getQueuedPrompts]; [cancelQueuedPrompt]
-  /// removes from it like a real queue owner and records the call.
-  final List<PluginQueuedPrompt> queuedPrompts = [];
-  final List<({String sessionId, String promptId})> cancelQueuedPromptCalls = [];
-
-  @override
-  Future<List<PluginQueuedPrompt>> getQueuedPrompts({required String sessionId}) async =>
-      List.unmodifiable(queuedPrompts);
-
-  @override
-  Future<bool> cancelQueuedPrompt({required String sessionId, required String promptId}) async {
-    cancelQueuedPromptCalls.add((sessionId: sessionId, promptId: promptId));
-    final index = queuedPrompts.indexWhere((prompt) => prompt.id == promptId);
-    if (index == -1) return false;
-    queuedPrompts.removeAt(index);
-    return true;
-  }
-
-  final _controller = StreamController<BridgeSseEvent>.broadcast();
-
-  // ── Configurable return values ───────────────────────────────────────────
-
-  List<PluginProject> projectsResult = [];
-  List<PluginSession> sessionsResult = [];
-  List<PluginCommand> commandsResult = [];
-  List<PluginMessageWithParts> messagesResult = [];
-  PluginProvidersResult providersResult = const PluginProvidersResult(providers: []);
-  PluginSession? createSessionResult;
-  PluginSession? renameSessionResult;
-  PluginProject? renameProjectResult;
-  List<PluginSession> childSessionsResult = [];
-  Map<String, PluginSessionStatus> sessionStatusesResult = {};
-  List<PluginAgent> agentsResult = [];
-  String? lastAgentsProjectId;
-  List<PluginPendingQuestion> pendingQuestionsResult = [];
-  List<PluginPendingPermission> pendingPermissionsResult = [];
-  PluginProject? currentProjectResult;
-
-  // ── Recorded call arguments ──────────────────────────────────────────────
-
-  String? lastGetSessionsWorktree;
-  int? lastGetSessionsStart;
-  int? lastGetSessionsLimit;
-  String? lastGetCommandsProjectId;
-
-  String? lastGetMessagesSessionId;
-  String? lastGetPendingQuestionsSessionId;
-  String? lastGetPendingPermissionsSessionId;
-
-  String? lastGetProvidersProjectId;
-  String? lastCreateSessionDirectory;
-  String? lastCreateSessionParentId;
-  String? lastCreateSessionProjectId;
-  List<PluginPromptPart>? lastCreateSessionParts;
-  String? lastCreateSessionUserVisibleText;
-  String? lastCreateSessionVariant;
-  String? lastCreateSessionAgent;
-  ({String providerID, String modelID})? lastCreateSessionModel;
-  String? lastRenameSessionId;
-  String? lastRenameSessionTitle;
-  String? lastRenameProjectId;
-  String? lastRenameProjectName;
-  String? lastDeleteSessionId;
-  String? lastArchiveSessionId;
-  String? lastDeleteWorkspaceProjectId;
-  String? lastDeleteWorkspaceWorktreePath;
-  String? lastGetChildSessionsSessionId;
-  String? lastSendPromptSessionId;
-  List<PluginPromptPart>? lastSendPromptParts;
-  String? lastSendPromptVariant;
-  String? lastSendPromptAgent;
-  ({String providerID, String modelID})? lastSendPromptModel;
-  String? lastSendCommandSessionId;
-  String? lastSendCommand;
-  String? lastSendCommandArguments;
-  String? lastSendCommandUserVisibleArguments;
-  String? lastSendCommandVariant;
-  String? lastSendCommandAgent;
-  ({String providerID, String modelID})? lastSendCommandModel;
-  String? lastAbortSessionId;
-  String? lastReplyQuestionId;
-  String? lastReplySessionId;
-  List<List<String>>? lastReplyAnswers;
-  String? lastRejectQuestionId;
-  String? lastRejectSessionId;
-  String? lastGetCurrentProjectProjectId;
-  String? lastReplyToPermissionRequestId;
-  String? lastReplyToPermissionSessionId;
-  PluginPermissionReply? lastReplyToPermissionReply;
-
-  // ── Error injection ──────────────────────────────────────────────────────
-
-  bool throwOnHealthCheck = false;
-  bool healthCheckResult = true;
-  int healthCheckCallCount = 0;
-  bool throwOnGetProjects = false;
-  Object? throwOnGetProjectsError;
-  Object? throwOnGetProjectError;
-  bool throwOnGetSessions = false;
-  Object? throwOnGetMessagesError;
-  Object? throwOnDeleteSessionError;
-  Object? throwOnArchiveSessionError;
-  Completer<void>? archiveSessionCompleter;
-  Completer<void>? sendCommandStarted;
-  Completer<void>? sendCommandCompleter;
-  Object? sendPromptError;
-  int getProjectsCallCount = 0;
-
-  // ── BridgePlugin implementation ──────────────────────────────────────────
-
-  @override
-  String get id => "fake";
-
-  @override
-  Stream<BridgeSseEvent> get events => _controller.stream;
-
-  void emitEvent(BridgeSseEvent event) => _controller.add(event);
-
-  Future<void> closeEvents() => _controller.close();
-
-  @override
-  Future<bool> healthCheck() async {
-    healthCheckCallCount++;
-    if (throwOnHealthCheck) throw Exception("healthCheck error");
-    return healthCheckResult;
-  }
-
-  @override
-  Future<List<PluginProject>> getProjects() async {
-    getProjectsCallCount++;
-    if (throwOnGetProjectsError case final error?) {
-      throw error;
-    }
-    if (throwOnGetProjects) throw Exception("getProjects error");
-    return projectsResult;
-  }
-
-  @override
-  Future<List<PluginSession>> getSessions(
-    String worktree, {
-    int? start,
-    int? limit,
-  }) async {
-    if (throwOnGetSessions) throw Exception("getSessions error");
-    lastGetSessionsWorktree = worktree;
-    lastGetSessionsStart = start;
-    lastGetSessionsLimit = limit;
-    return sessionsResult;
-  }
-
-  @override
-  Future<List<PluginCommand>> getCommands({required String? projectId}) async {
-    lastGetCommandsProjectId = projectId;
-    return commandsResult;
-  }
-
-  @override
-  Future<PluginSessionOptionsDiscoveryResult> getSessionOptions({
-    required String projectId,
-    required PluginSessionOptionsDiscoveryMode discoveryMode,
-  }) => throw UnimplementedError();
-
-  @override
-  Future<PluginSession> createSession({
-    required String directory,
-    required String? parentSessionId,
-    required List<PluginPromptPart> parts,
-    required String? userVisibleText,
-    required PluginSessionVariant? variant,
-    required String? agent,
-    required ({String providerID, String modelID})? model,
-  }) async {
-    lastCreateSessionDirectory = directory;
-    lastCreateSessionParentId = parentSessionId;
-    lastCreateSessionProjectId = directory;
-    lastCreateSessionParts = parts;
-    lastCreateSessionUserVisibleText = userVisibleText;
-    lastCreateSessionVariant = variant?.id;
-    lastCreateSessionAgent = agent;
-    lastCreateSessionModel = model;
-    return createSessionResult ??
-        const PluginSession(
-          id: "",
-          projectID: "",
-          directory: "",
-          parentID: null,
-          title: null,
-          time: null,
-        );
-  }
-
-  @override
-  Future<PluginSession> renameSession({
-    required String sessionId,
-    required String title,
-  }) async {
-    lastRenameSessionId = sessionId;
-    lastRenameSessionTitle = title;
-    return renameSessionResult ??
-        const PluginSession(
-          id: "",
-          projectID: "",
-          directory: "",
-          parentID: null,
-          title: null,
-          time: null,
-        );
-  }
-
-  @override
-  Future<PluginProject> renameProject({
-    required String projectId,
-    required String name,
-  }) async {
-    lastRenameProjectId = projectId;
-    lastRenameProjectName = name;
-    return renameProjectResult ?? const PluginProject(id: "", directory: "");
-  }
-
-  @override
-  Future<void> deleteSession(String sessionId) async {
-    lastDeleteSessionId = sessionId;
-    if (throwOnDeleteSessionError case final error?) {
-      throw error;
-    }
-  }
-
-  @override
-  Future<void> archiveSession({required String sessionId}) async {
-    lastArchiveSessionId = sessionId;
-    if (throwOnArchiveSessionError case final error?) {
-      throw error;
-    }
-    if (archiveSessionCompleter case final completer?) {
-      await completer.future;
-    }
-  }
-
-  @override
-  Future<void> deleteWorkspace({
-    required String projectId,
-    required String worktreePath,
-  }) async {
-    lastDeleteWorkspaceProjectId = projectId;
-    lastDeleteWorkspaceWorktreePath = worktreePath;
-  }
-
-  @override
-  Future<List<PluginSession>> getChildSessions(String sessionId) async {
-    lastGetChildSessionsSessionId = sessionId;
-    return childSessionsResult;
-  }
-
-  @override
-  Future<Map<String, PluginSessionStatus>> getSessionStatuses() async => sessionStatusesResult;
-
-  @override
-  Future<List<PluginMessageWithParts>> getSessionMessages(
-    String sessionId,
-  ) async {
-    lastGetMessagesSessionId = sessionId;
-    if (throwOnGetMessagesError case final error?) {
-      throw error;
-    }
-    return messagesResult;
-  }
-
-  @override
-  Future<void> sendPrompt({
-    required String promptId,
-    required String sessionId,
-    required List<PluginPromptPart> parts,
-    required PluginSessionVariant? variant,
-    required String? agent,
-    required ({String providerID, String modelID})? model,
-  }) async {
-    if (sendPromptError case final error?) throw error;
-    lastSendPromptSessionId = sessionId;
-    lastSendPromptParts = parts;
-    lastSendPromptVariant = variant?.id;
-    lastSendPromptAgent = agent;
-    lastSendPromptModel = model;
-  }
-
-  @override
-  Future<void> sendCommand({
-    required String promptId,
-    required String sessionId,
-    required String command,
-    required String arguments,
-    required String? userVisibleArguments,
-    required PluginSessionVariant? variant,
-    required String? agent,
-    required ({String providerID, String modelID})? model,
-  }) async {
-    if (sendCommandStarted case final started? when !started.isCompleted) started.complete();
-    lastSendCommandSessionId = sessionId;
-    lastSendCommand = command;
-    lastSendCommandArguments = arguments;
-    lastSendCommandUserVisibleArguments = userVisibleArguments;
-    lastSendCommandVariant = variant?.id;
-    lastSendCommandAgent = agent;
-    lastSendCommandModel = model;
-    if (sendCommandCompleter case final completer?) {
-      await completer.future;
-    }
-  }
-
-  @override
-  Future<void> abortSession({required String sessionId}) async {
-    lastAbortSessionId = sessionId;
-  }
-
-  @override
-  Future<List<PluginAgent>> getAgents({required String projectId}) async {
-    lastAgentsProjectId = projectId;
-    return agentsResult;
-  }
-
-  @override
-  Future<List<PluginPendingQuestion>> getPendingQuestions({required String sessionId}) async {
-    lastGetPendingQuestionsSessionId = sessionId;
-    return pendingQuestionsResult;
-  }
-
-  @override
-  Future<List<PluginPendingQuestion>> getProjectQuestions({required String projectId}) async => pendingQuestionsResult;
-
-  @override
-  Future<List<PluginPendingPermission>> getPendingPermissions({required String sessionId}) async {
-    lastGetPendingPermissionsSessionId = sessionId;
-    return pendingPermissionsResult;
-  }
-
-  @override
-  Future<void> replyToQuestion({
-    required String questionId,
-    required String sessionId,
-    required List<List<String>> answers,
-  }) async {
-    lastReplyQuestionId = questionId;
-    lastReplySessionId = sessionId;
-    lastReplyAnswers = answers;
-  }
-
-  @override
-  Future<void> rejectQuestion({required String questionId, required String? sessionId}) async {
-    lastRejectQuestionId = questionId;
-    lastRejectSessionId = sessionId;
-  }
-
-  @override
-  Future<void> replyToPermission({
-    required String requestId,
-    required String sessionId,
-    required PluginPermissionReply reply,
-  }) async {
-    lastReplyToPermissionRequestId = requestId;
-    lastReplyToPermissionSessionId = sessionId;
-    lastReplyToPermissionReply = reply;
-  }
-
-  @override
-  Future<PluginProject> getProject(String projectId) async {
-    if (throwOnGetProjectError case final error?) {
-      throw error;
-    }
-    lastGetCurrentProjectProjectId = projectId;
-    return currentProjectResult ?? const PluginProject(id: "", directory: "");
-  }
-
-  @override
-  List<PluginProjectActivitySummary> getActiveSessionsSummary() => [];
-
-  @override
-  Future<PluginProvidersResult> getProviders({required String projectId}) async {
-    lastGetProvidersProjectId = projectId;
-    return providersResult;
-  }
-
-  @override
-  Future<void> dispose() async {}
-
-  Future<void> close() => _controller.close();
 }
 
 /// Hand-written fake [SessionDao] for testing.
@@ -587,109 +203,6 @@ class FakeSessionDao() {
   Future<void> deleteSession({required String sessionId}) async {
     _sessions.remove(sessionId);
   }
-}
-
-/// Hand-written fake [SessionMetadataRepository] for testing.
-class FakeSessionMetadataRepository() implements SessionMetadataRepository {
-  String? generateResult = "Generated title";
-  String generatedBranchName = "generated-branch";
-  String? lastGenerateMessage;
-
-  @override
-  void beginShutdown() {}
-
-  @override
-  Future<GeneratedSessionMetadata> generateMetadata({required String firstMessage}) async {
-    lastGenerateMessage = firstMessage;
-    return (
-      title: generateResult ?? (throw StateError("metadata unavailable")),
-      branchName: generatedBranchName,
-    );
-  }
-}
-
-class FakePullRequestRepository() implements PullRequestRepository {
-  final Map<String, List<PullRequestDto>> _prsBySessionId = <String, List<PullRequestDto>>{};
-  final Map<String, PullRequestDto> _prsByPrimaryKey = <String, PullRequestDto>{};
-
-  void setPr({required String sessionId, required PullRequestDto pullRequest}) {
-    _prsBySessionId.putIfAbsent(sessionId, () => <PullRequestDto>[]).add(pullRequest);
-    _prsByPrimaryKey[_key(
-          projectId: pullRequest.projectId,
-          githubRepositoryIdentity: pullRequest.githubRepositoryIdentity,
-          prNumber: pullRequest.prNumber,
-        )] =
-        pullRequest;
-  }
-
-  Future<Map<String, List<PullRequestDto>>> getPrsBySessionIds({
-    required List<String> sessionIds,
-    required VerifiedGithubLogin verifiedGithubLogin,
-  }) async {
-    return <String, List<PullRequestDto>>{
-      for (final sessionId in sessionIds)
-        if (_prsBySessionId[sessionId]?.where((pr) => pr.githubLogin == verifiedGithubLogin.login).toList()
-            case final matching? when matching.isNotEmpty)
-          sessionId: matching,
-    };
-  }
-
-  @override
-  Future<PullRequestReplacementOutcome> replaceScopedPullRequests({
-    required String projectId,
-    required VerifiedGithubLogin verifiedGithubLogin,
-    required Map<String, String> capturedRootDirectoriesBySessionId,
-    required List<PullRequestTargetSelection> targetSelections,
-    required int lastCheckedAt,
-  }) async {
-    _prsByPrimaryKey.removeWhere((_, pullRequest) => pullRequest.projectId == projectId);
-    for (final selection in targetSelections) {
-      if (selection is! PullRequestTargetSelected) continue;
-      final pullRequest = selection;
-      final record = PullRequestDto(
-        projectId: projectId,
-        githubRepositoryIdentity: pullRequest.target.githubRepositoryIdentity,
-        githubLogin: verifiedGithubLogin.login,
-        prNumber: pullRequest.number,
-        branchName: pullRequest.target.branchName,
-        url: pullRequest.url,
-        title: pullRequest.title,
-        state: pullRequest.state,
-        mergeableStatus: pullRequest.mergeableStatus,
-        reviewDecision: pullRequest.reviewDecision,
-        checkStatus: pullRequest.checkStatus,
-        lastCheckedAt: lastCheckedAt,
-        createdAt: pullRequest.createdAt.millisecondsSinceEpoch,
-      );
-      _prsByPrimaryKey[_key(
-            projectId: record.projectId,
-            githubRepositoryIdentity: record.githubRepositoryIdentity,
-            prNumber: record.prNumber,
-          )] =
-          record;
-    }
-    return const PullRequestReplacementApplied(changed: true);
-  }
-
-  String _key({
-    required String projectId,
-    required String githubRepositoryIdentity,
-    required int prNumber,
-  }) {
-    return "$projectId::$githubRepositoryIdentity::$prNumber";
-  }
-
-  @override
-  Future<Set<String>> prepareScopedRefresh({
-    required Set<String> projectIds,
-    required VerifiedGithubLogin verifiedGithubLogin,
-  }) async => const <String>{};
-
-  @override
-  Future<Set<String>> applyResolvedTargets({
-    required Map<String, List<StoredSession>> sessionsByProject,
-    required Map<String, PullRequestDirectoryTarget> targetsByDirectory,
-  }) async => const <String>{};
 }
 
 class FakePrSyncService({
